day12..... wasted lots of time with a small bug (+ instead of -)

+++ challenges/internal/monkeys/monkey_test.go
@@ -12,23 +12,23 @@ func TestParseMonkeyNumber(t *testing.T) {
 }
 
 func TestParseMonkeyItems(t *testing.T) {
-	assert.Equal(t, []int{79, 98}, parseMonkeyItems("  Starting items: 79, 98"))
-	assert.Equal(t, []int{54, 65, 75, 74}, parseMonkeyItems("Starting items: 54, 65, 75, 74"))
+	assert.Equal(t, []int64{79, 98}, parseMonkeyItems("  Starting items: 79, 98"))
+	assert.Equal(t, []int64{54, 65, 75, 74}, parseMonkeyItems("Starting items: 54, 65, 75, 74"))
 }
 
 func TestParseMonkeyOperation(t *testing.T) {
 	f := parseMonkeyOperation("  Operation: new = old * 19")
-	assert.Equal(t, 38, f(2))
+	assert.Equal(t, int64(38), f(2))
 
 	f = parseMonkeyOperation("  Operation: new = old * old")
-	assert.Equal(t, 4, f(2))
+	assert.Equal(t, int64(4), f(2))
 
 	f = parseMonkeyOperation("  Operation: new = old + 123")
-	assert.Equal(t, 125, f(2))
+	assert.Equal(t, int64(125), f(2))
 }
 
 func TestParseMonkeyTest(t *testing.T) {
-	assert.Equal(t, 19, parseMonkeyTest("  Test: divisible by 19"))
+	assert.Equal(t, int64(19), parseMonkeyTest("  Test: divisible by 19"))
 }
 
 func TestParseMonkeyAction(t *testing.T) {
